pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "fdtdw"
version = "0.1.0"
description = "fdtd package"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "numpy",
    "warp-lang",
    "pyvista",
    "imageio",
]

[project.optional-dependencies]
dev = [
    "pytest",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["fdtdw*"]
